requirements.txt
flask
flask-cors
python-dotenv
google-generativeai
scapy
wifi